Add %LEN and fixed DSPLY
Took me a while to get LEN to work. Not had to use the instance keyword
in CIL much before, so I had a lot of trial and error. At first, I made
%LEN it's own method in the CIL and it would just return the length of
the string passed in - but then I figured out what was actually
happening and got rid of it.. making it cleaner.

Also, you can pass in expressions to DSPLY and it will work as expected
whereas before it would only take the first piece of the expression.