Array implementation
Arrays are ready to be implemented.
To finish them off, inside of Module.cs we must add the Dim keyword to
Dcl-S, change the callProc method inside of Procedure.cs to load the
index element of an array onto the stack - and then it should be
finished.

Also, to do all of this I had to create a Variable.cs class.. so we can
store information about Variables now.